Rename keys and provider files
